requirements.txt
requests==2.31.0
aiogram==3.4.1
python-dotenv==1.0.0
aiosqlite==0.19.0